AlphaQCM/requirements_crypto.txt
ccxt>=4.0.0
torch>=2.0.0
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
tensorboard>=2.10.0
gymnasium>=1.0.0
